refactor: split out type guards, rename vars
