+++ packages/caip/src/index.ts
@@ -3,3 +3,5 @@ export * from './chainId/chainId'
 export * from './accountId/accountId'
 export * from './assetId/assetId'
 export * from './utils'
+export * from './constants'
+export * from './typeGuards'
